requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
bidict==0.22.1
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
cmake==3.30.4
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
docker-pycreds==0.4.0
exceptiongroup==1.2.2
fastapi==0.115.2
ffmpy==0.4.0
filelock==3.9.0
fonttools==4.39.2
fsspec==2024.9.0
gitdb==4.0.10
GitPython==3.1.31
gradio==5.1.0
gradio_client==1.4.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
jsonpickle==3.0.1
kiwisolver==1.4.4
lit==18.1.8
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
miditoolkit==0.1.16
mido==1.2.10
more-itertools==9.1.0
mpmath==1.2.1
music21==8.1.0
muspy==0.5.0
networkx==3.0
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
orjson==3.10.7
packaging==23.0
pandas==2.2.3
pathtools==0.1.2
Pillow==9.3.0
pretty-midi==0.2.10
protobuf==4.22.1
psutil==5.9.4
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.0.9
pypianoroll==1.0.4
python-dateutil==2.8.2
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0
requests==2.28.1
rich==13.9.2
ruff==0.6.9
scipy==1.10.1
semantic-version==2.10.0
sentry-sdk==1.18.0
setproctitle==1.3.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
starlette==0.40.0
sympy==1.11.1
tomlkit==0.12.0
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tqdm==4.65.0
triton==2.0.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.13
uvicorn==0.32.0
wandb==0.14.0
webcolors==1.13
websockets==12.0
